+++ azext_iot/central/_help.py
@@ -86,6 +86,35 @@ def _load_central_devices_help():
     """
 
     helps[
+        "iot central device manual-failover"
+    ] = """
+        type: command
+        short-summary: Execute a manual failover of device across multiple IoT Hubs to validate device firmware's ability to reconnect using DPS to a different IoT Hub.
+
+        examples:
+        - name: Execute a manual failover of device across multiple IoT Hubs to validate device firmware's ability to reconnect using DPS to a different IoT Hub.
+          text: >
+            az iot central device manual-failover
+            --app-id {appid}
+            --device-id {deviceid}
+            --ttl-minutes {ttl_minutes}
+    """
+
+    helps[
+        "iot central device manual-failback"
+    ] = """
+        type: command
+        short-summary: Reverts the previously executed failover command by moving the device back to it's original IoT Hub
+
+        examples:
+        - name: Reverts the previously executed failover command by moving the device back to it's original IoT Hub
+          text: >
+            az iot central device manual-failback
+            --app-id {appid}
+            --device-id {deviceid}
+    """
+    
+    helps[
         "iot central device delete"
     ] = """
         type: command